backend/web-scraping-service/requirements.txt
fastapi
uvicorn
newspaper3k
requests
python-dotenv
